requirements.txt
pandas==2.0.0
openpyxl==3.1.2
qrcode==7.4.2
pillow==10.0.0